Enum teaclave_rpc::Code

pub enum Code {

    Ok,
    Cancelled,
    Unknown,
    InvalidArgument,
    DeadlineExceeded,
    NotFound,
    AlreadyExists,
    PermissionDenied,
    ResourceExhausted,
    FailedPrecondition,
    Aborted,
    OutOfRange,
    Unimplemented,
    Internal,
    Unavailable,
    DataLoss,
    Unauthenticated,

}

gRPC status codes used by Status.

These variants match the gRPC status codes.

Variants

Ok

The operation completed successfully.

Cancelled

The operation was cancelled.

Unknown

Unknown error.

InvalidArgument

Client specified an invalid argument.

DeadlineExceeded

Deadline expired before operation could complete.

NotFound

Some requested entity was not found.

AlreadyExists

Some entity that we attempted to create already exists.

PermissionDenied

The caller does not have permission to execute the specified operation.

ResourceExhausted

Some resource has been exhausted.

FailedPrecondition

The system is not in a state required for the operation’s execution.

Aborted

The operation was aborted.

OutOfRange

Operation was attempted past the valid range.

Unimplemented

Operation is not implemented or not supported.

Internal

Internal error.

Unavailable

The service is currently unavailable.

DataLoss

Unrecoverable data loss or corruption.

Unauthenticated

The request does not have valid authentication credentials

Implementations

impl Code

pub fn description(&self) -> &'static str

Get description of this Code.

fn make_grpc_request() -> tonic::Code {
    // ...
    tonic::Code::Ok
}
let code = make_grpc_request();
println!("Operation completed. Human readable description: {}", code.description());

If you only need description in println, format, log and other formatting contexts, you may want to use Display impl for Code instead.

impl Code

pub fn from_i32(i: i32) -> Code

Get the Code that represents the integer, if known.

If not known, returns Code::Unknown (surprise!).

pub fn from_bytes(bytes: &[u8]) -> Code

Convert the string representation of a Code (as stored, for example, in the grpc-status header in a response) into a Code. Returns Code::Unknown if the code string is not a valid gRPC status code.